pyproject.toml
[project]
name = "flask-file-explorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=2.0",
    "opencv-python>=4.7",
    "piexif>=1.1.3",
    "pillow>=9.0",
    "pillow-heif>=0.7.0",
    "python-dotenv>=0.20",
    "python-magic>=0.4.27",
    "tqdm>=4.0",
]
